subject crud updated

+++ src/containers/SubjectDetailView.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import axios from 'axios';
-
 import { Button, Card } from 'antd';
-
+import { connect } from 'react-redux';
 import CustomForm from '../components/Form';
 
 class SubjectDetail extends React.Component {
@@ -11,21 +10,38 @@ class SubjectDetail extends React.Component {
         subject: {}
     }
 
-    componentDidMount() {
-        const classID = this.props.match.params.classID;
-        axios.get(`https://diploma.zharaskhan.com/api/classes/${classID}/`)
-            .then(res => {
-                this.setState({
-                    subject: res.data
-                });
-            })
+    componentWillReceiveProps(newProps){
+        console.log(newProps);
+        if(newProps.token){
+            axios.defaults.headers = {
+                "Content-Type": "application/json",
+                Authorization: newProps.token
+            }
+            const classID = this.props.match.params.classID;
+            axios.get(`https://diploma.zharaskhan.com/api/classes/${classID}/`)
+                .then(res => {
+                    this.setState({
+                        subject: res.data
+                    });
+                })
+        }
     }
 
+
     handleDelete = (event) => {
-        const classID = this.props.match.params.classID;
-        axios.delete(`https://diploma.zharaskhan.com/api/classes/${classID}/`);
-        this.props.history.push('/');
-        this.forceUpdate();
+        if(this.props.token !== null){
+            const classID = this.props.match.params.classID;
+            axios.defaults.headers = {
+                "Content-Type": "application/json",
+                Authorization: this.props.token
+            }
+            axios.delete(`https://diploma.zharaskhan.com/api/classes/${classID}/`);
+            this.props.history.push('/');
+            this.forceUpdate();
+        }else{
+            //some kind of message
+        }
+
     }
 
     render() {
@@ -45,5 +61,9 @@ class SubjectDetail extends React.Component {
         )
     }
 }
-
-export default SubjectDetail;
+const mapStateToProps = state => {
+    return {
+      token: state.token
+    }
+  }
+export default connect(mapStateToProps)(SubjectDetail);
